@@ -1,9 +1,9 @@
-﻿using System;
-
-namespace Generator.Core.Metamodel
-{
-	public class BaseModel : IMetamodelNode
-	{
-		public Guid Id { get; set; }
-	}
-}
+﻿// using System;
+//
+// namespace Generator.Core.Metamodel
+// {
+// 	public class BaseModel : IMetamodelNode
+// 	{
+// 		public Guid Id { get; set; }
+// 	}
+// }
